requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pygame-ce==2.5.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
